LayoutTests/accessibility/element-haspopup-expected.txt

Link 1 Link 2
Paragraph 1.

Paragraph 2.

Paragraph 3.

This tests that an element will report that it has a popup correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.accessibleElementById('link1').hasPopup is true
PASS accessibilityController.accessibleElementById('link2').hasPopup is false
PASS accessibilityController.accessibleElementById('paragraph1').hasPopup is true
PASS accessibilityController.accessibleElementById('paragraph2').hasPopup is false
PASS accessibilityController.accessibleElementById('paragraph3').hasPopup is false
PASS successfullyParsed is true

TEST COMPLETE
